+++ server/benchmarks/bench.sh
@@ -13,7 +13,7 @@ cat <<EOL
 
 Run hasura benchmarks
 
-Usage:   
+Usage:
     $ $0 <benchmark_dir> [<hasura_docker_image>] [<sleep_time_sec_before_bench>]
 
 The first argument chooses the particular benchmark set to run e.g. "chinook"
@@ -29,7 +29,7 @@ exit 1
 }
 
 [ ! -d "benchmark_sets/${1-}" ] && die_usage
-BENCH_DIR="benchmark_sets/$1"
+BENCH_DIR="$(pwd)/benchmark_sets/$1"
 REQUESTED_HASURA_DOCKER_IMAGE="${2-}"
 # We may wish to sleep after setting up the schema, etc. to e.g. allow memory
 # to settle to a baseline before we measure it:
@@ -47,8 +47,16 @@ function cleanup {
       || echo "Stopping hasura failed, maybe it never started?"
   fi
   pg_cleanup || echo "Stopping postgres failed, maybe it never started?"
+
+  custom_cleanup || echo "Custom cleanup failed"
 }
 trap cleanup EXIT
+
+if [ $(uname -s) = Darwin ]; then
+  DOCKER_LOCALHOST=host.docker.internal
+else
+  DOCKER_LOCALHOST=127.0.0.1
+fi
 
 # The beefy c4.8xlarge EC2 instance has two sockets, so we'll try our best to
 # pin hasura on one and postgres on the other
@@ -157,7 +165,7 @@ CONF_FLAGS=$(echo "$CONF" | sed  -e 's/^/-c /'  | tr '\n' ' ')
 #   numbers we get here are useful in absolute terms as well, representing ideal
 #   performance)
 #
-# [1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSPerformance.html 
+# [1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSPerformance.html
 # [2]: https://performance.sunlight.io/postgres/
 function pg_launch_container(){
   echo_pretty "Launching postgres container: $PG_CONTAINER_NAME"
@@ -188,10 +196,11 @@ function pg_cleanup(){
 ######################
 #   graphql-engine   #
 ######################
-  
+
 # This matches the default we use in `dev.sh graphql-engine`
 HASURA_GRAPHQL_SERVER_PORT=8181
 HASURA_URL="http://127.0.0.1:$HASURA_GRAPHQL_SERVER_PORT"
+HASURA_DOCKER_URL="http://$DOCKER_LOCALHOST:$HASURA_GRAPHQL_SERVER_PORT"
 
 # Maybe launch the hasura instance we'll benchmark
 function maybe_launch_hasura_container() {
@@ -214,7 +223,7 @@ function maybe_launch_hasura_container() {
 
 function hasura_wait() {
   # Wait for the graphql-engine under bench to be ready
-  echo -n "Waiting for graphql-engine"
+  echo -n "Waiting for graphql-engine at $HASURA_URL"
   until curl -s "$HASURA_URL/v1/query" &>/dev/null; do
     echo -n '.' && sleep 0.2
   done
@@ -235,12 +244,12 @@ function install_latest_graphql_bench() {
   echo_pretty "Installing/updating graphql-bench"
   graphql_bench_git=$(mktemp -d -t graphql-bench-XXXXXXXXXX)
   git clone --depth=1 https://github.com/hasura/graphql-bench.git "$graphql_bench_git"
-  
+
   cd "$graphql_bench_git"
   # We name this 'graphql-bench-ci' so it doesn't interfere with other versions
   # (e.g. local dev of `graphql-bench`, installed with `make
   # build_local_docker_image`:
-  docker build -t graphql-bench-ci:latest ./app 
+  docker build -t graphql-bench-ci:latest ./app
   cd -
   echo_pretty "Done"
 }
@@ -254,19 +263,49 @@ function run_benchmarks() {
   $TASKSET_K6 docker run --net=host -v "$PWD":/app/tmp -i $K6_DOCKER_t_OR_init \
     graphql-bench-ci query \
     --config="./tmp/config.query.yaml" \
-    --outfile="./tmp/report.json" --url "$HASURA_URL/v1/graphql"
+    --outfile="./tmp/report.json" --url "$HASURA_DOCKER_URL/v1/graphql"
 
   echo_pretty "Done. Report at $PWD/report.json"
+  cd -
+}
+
+function custom_setup() {
+  cd "$BENCH_DIR"
+  if [ -x setup.sh ]; then
+    echo_pretty "Running custom setup script"
+    ./setup.sh
+  fi
+
+  cd -
+}
+
+function custom_cleanup() {
+  cd "$BENCH_DIR"
+  if [ -x cleanup.sh ]; then
+    echo_pretty "Running custom cleanup script"
+    ./cleanup.sh
+  fi
+
   cd -
 }
 
 function load_data_and_schema() {
   echo_pretty "Loading data and adding schema"
   cd "$BENCH_DIR"
-  gunzip -c dump.sql.gz | $PSQL_DOCKER &> /dev/null
-  # --fail-with-body is what we want, but is not available on older curl:
-  # TODO LATER: use /v1/metadata once stable
-  curl --fail -X POST -H "Content-Type: application/json" -d @replace_metadata.json "$HASURA_URL/v1/query"
+  if [ -f dump.sql.gz ]; then
+    gunzip -c dump.sql.gz | $PSQL_DOCKER &> /dev/null
+  else
+    echo_pretty "No data to load"
+  fi
+
+  if [ -f replace_metadata.json ]; then
+    # --fail-with-body is what we want, but is not available on older curl:
+    # TODO LATER: use /v1/metadata once stable
+    curl --fail -X POST -H "Content-Type: application/json" -d @replace_metadata.json "$HASURA_URL/v1/query"
+  else
+    echo_pretty "No metadata to replace"
+  fi
+
   cd -
 }
 
@@ -283,6 +322,8 @@ install_latest_graphql_bench
 pg_wait
 maybe_launch_hasura_container
 hasura_wait
+
+custom_setup
 
 load_data_and_schema
 run_benchmarks