+++ server/src-lib/Hasura/GraphQL/Transport/HTTP/Protocol.hs
@@ -5,6 +5,8 @@ module Hasura.GraphQL.Transport.HTTP.Protocol
   , GQLReqParsed
   , GQLReqOutgoing
   , renderGQLReqOutgoing
+  , SingleOperation
+  , getSingleOperation
   , toParsed
   , GQLQueryText(..)
   , GQLExecDoc(..)
@@ -20,6 +22,7 @@ module Hasura.GraphQL.Transport.HTTP.Protocol
   , isExecError
   ) where
 
+import           Data.Text.Extended             (dquote)
 import           Hasura.Prelude
 
 import qualified Data.Aeson                     as J
@@ -27,6 +30,7 @@ import qualified Data.Aeson.Casing              as J
 import qualified Data.Aeson.TH                  as J
 import qualified Data.ByteString.Lazy           as BL
 import qualified Data.HashMap.Strict            as Map
+import qualified Hasura.GraphQL.Execute.Inline  as EI
 import qualified Language.GraphQL.Draft.Parser  as G
 import qualified Language.GraphQL.Draft.Printer as G
 import qualified Language.GraphQL.Draft.Syntax  as G
@@ -119,7 +123,12 @@ type GQLReqParsed = GQLReq GQLExecDoc
 -- '_todName' if present.
 --
 -- These could maybe benefit from an HKD refactoring.
-type GQLReqOutgoing = GQLReq (G.TypedOperationDefinition G.NoFragments G.Name)
+type GQLReqOutgoing = GQLReq SingleOperation
+
+-- | A single graphql operation to be executed, with fragment definitions
+-- inlined. This is the simplified form of 'GQLExecDoc' or
+-- 'G.ExecutableDocument':
+type SingleOperation = G.TypedOperationDefinition G.NoFragments G.Name
 
 renderGQLReqOutgoing :: GQLReqOutgoing -> GQLReqUnparsed
 renderGQLReqOutgoing = fmap (GQLQueryText . G.renderExecutableDoc . toExecDoc . inlineFrags)
@@ -131,6 +140,40 @@ renderGQLReqOutgoing = fmap (GQLQueryText . G.renderExecutableDoc . toExecDoc . 
       opDef { G._todSelectionSet = G.fmapSelectionSetFragment G.inline $ G._todSelectionSet opDef }
     toExecDoc =
       G.ExecutableDocument . pure . G.ExecutableDefinitionOperation . G.OperationDefinitionTyped
+
+-- | Obtain the actual single operation to be executed, from the possibly-
+-- multi-operation document, validating per the spec and inlining any
+-- fragment definitions (pre-defined parts of a graphql query) at fragment
+-- spreads (locations where fragments are "spliced"). See:
+--
+--     https://spec.graphql.org/June2018/#sec-Executable-Definitions  and...
+--     https://graphql.org/learn/serving-over-http/
+getSingleOperation
+  :: MonadError QErr m
+  => GQLReqParsed
+  -> m SingleOperation
+getSingleOperation (GQLReq opNameM q _varValsM) = do
+  let (selSets, opDefs, fragments) = G.partitionExDefs $ unGQLExecDoc q
+  G.TypedOperationDefinition{..} <-
+    case (opNameM, selSets, opDefs) of
+      (Just opName, [], _) -> do
+        let n = _unOperationName opName
+            opDefM = find (\opDef -> G._todName opDef == Just n) opDefs
+        onNothing opDefM $ throw400 ValidationFailed $
+          "no such operation found in the document: " <> dquote n
+      (Just _, _, _)  ->
+        throw400 ValidationFailed $ "operationName cannot be used when " <>
+        "an anonymous operation exists in the document"
+      (Nothing, [selSet], []) ->
+        return $ G.TypedOperationDefinition G.OperationTypeQuery Nothing [] [] selSet
+      (Nothing, [], [opDef])  ->
+        return opDef
+      (Nothing, _, _) ->
+        throw400 ValidationFailed $ "exactly one operation has to be present " <>
+        "in the document when operationName is not specified"
+
+  inlinedSelSet <- EI.inlineSelectionSet fragments _todSelectionSet
+  pure $ G.TypedOperationDefinition{_todSelectionSet = inlinedSelSet, ..}
 
 toParsed :: (MonadError QErr m ) => GQLReqUnparsed -> m GQLReqParsed
 toParsed req = case G.parseExecutableDoc gqlText of