+++ server/src-lib/Hasura/Server/Utils.hs
@@ -160,8 +160,11 @@ commonResponseHeadersIgnored =
   , "Content-Type", "Content-Length"
   ]
 
+sessionVariablePrefix :: Text
+sessionVariablePrefix = "x-hasura-"
+
 isSessionVariable :: Text -> Bool
-isSessionVariable = T.isPrefixOf "x-hasura-" . T.toLower
+isSessionVariable = T.isPrefixOf sessionVariablePrefix . T.toLower
 
 isReqUserId :: Text -> Bool
 isReqUserId = (== "req_user_id") . T.toLower