+++ server/tests-py/test_remote_schema_permissions.py
@@ -127,6 +127,23 @@ class TestRemoteSchemaPermissionsExecution:
         check_query_f(hge_ctx, self.dir() + 'unknown_role_execution.yaml')
 
 @use_test_fixtures
+class TestCustomizedRemoteSchemaPermissionsExecution:
+
+    @classmethod
+    def dir(cls):
+        return "queries/remote_schemas/permissions/schema_customization/"
+
+    def test_execution_with_subset_of_fields_exposed_to_role(self, hge_ctx):
+        st_code, resp = hge_ctx.v1metadataq_f(self.dir() + 'add_permission_with_valid_subset_of_fields.yaml')
+        assert st_code == 200, resp
+        check_query_f(hge_ctx, self.dir() + 'execution_with_partial_fields_exposed_to_role.yaml')
+
+    def test_execution_with_subset_of_arguments_exposed_to_role(self, hge_ctx):
+        st_code, resp = hge_ctx.v1metadataq_f(self.dir() + 'add_permission_with_valid_subset_of_arguments.yaml')
+        assert st_code == 200, resp
+        check_query_f(hge_ctx, self.dir() + 'execution_with_partial_args_exposed_to_role.yaml')
+
+@use_test_fixtures
 class TestRemoteSchemaPermissionsArgumentPresets:
 
     @classmethod